+++ packages/jac/src/index.ts
@@ -1,4 +1,3 @@
-export * from '@/enums'
 export * from '@/errors'
 export * from '@/helpers'
 export * from '@/json-api'