+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `@distributedlab/jac` - `@distributedlab/fetcher` dependency
+
+### Removed
+- `@distributedlab/jac` - `axios` dependency
+
 ## [0.2.0-rc.0] - 2023-04-05
 ### Added
 - `root` - "Using in the projects made by create-react-app" section in the `README.md`